+++ app/src/test/java/com/techcourse/service/UserServiceTest.java
@@ -36,7 +36,7 @@ class UserServiceTest {
     @Test
     void testChangePassword() {
         final var userHistoryDao = new UserHistoryDao(jdbcTemplate);
-        final var userService = new UserService(userDao, userHistoryDao, transactionTemplate);
+        final var userService = new TxUserService(new AppUserService(userDao, userHistoryDao), transactionTemplate);
 
         final var newPassword = "qqqqq";
         final var createBy = "gugu";
@@ -51,7 +51,7 @@ class UserServiceTest {
     void testTransactionRollback() {
         // 트랜잭션 롤백 테스트를 위해 mock으로 교체
         final var userHistoryDao = new MockUserHistoryDao(jdbcTemplate);
-        final var userService = new UserService(userDao, userHistoryDao, transactionTemplate);
+        final var userService = new TxUserService(new AppUserService(userDao, userHistoryDao), transactionTemplate);
 
         final var newPassword = "newPassword";
         final var createBy = "gugu";
